kolejna

+++ main.py
@@ -33,4 +33,4 @@ for promien_bok in liczby:
   if len(promien_bok)>3:
     print("Błąd: można podać maksymalnie 3 liczby")
    
-print(round(wyniki,2))
+print(int(round(wyniki,2)))
